requirements.txt
python-telegram-bot[all]==20.7
pyrogram==2.0.106
tgcrypto==1.2.5
aiosqlite==0.20.0
jieba==0.42.1
python-dotenv==1.0.0
apscheduler==3.10.4
